DETECTING COMPUTER VISION SYNDROME-Project Report
We Spend Most Of Our Time In Front Of Digital Devices, Be It At Home, At Work
Or While Travelling. When Working With Screens Is Important To Blink Regularly, Failure Of Which Can Result In A Condition Known As Computer Vision
Syndrome(CVS). To Avoid The Above Situation, We Can Create A System Which
Involves Blink Detection And If The Rate Falls Below The Normal Threshold The
Person Is Warned Immediately. This Paper Overviews Different Algorithms Used In
Blink Detection. The Various Phases In This System Are Image Capture, Face Detection, Eye Localization And Finally Blink Detection. The Standard Algorithms
For Each Phase Have Been Discussed. We Have Proposed A Method For Blink
Detection Using Convolutional Neural Network To Detect The Eye States And Predict The Blinks. Finally, A Comparison Of Our Method Is Made To The Other
Methods Of Detecting Blinks In Both Still Images And Live Video Feed Scenarios.
Regular Usage Of The Proposed System Could Significantly Reduce The Signs Of
Computer Vision Syndrome And Ensure Efficient Work Experience For People With
Long Term Computer Use.

This is the report of "DETECTING COMPUTER VISION SYNDROME" project.